requirements.txt
# Requirements for TARGET MACHINE (Windows 7 32-bit with Python 3.8.10)

# Cross-platform keylogger (Python 3.8 compatible)
pynput>=1.7.3,<=1.7.6

# Audio recording (Python 3.8 compatible) 
PyAudio>=0.2.11

# Video recording and screenshots (Python 3.8 compatible)
opencv-python>=4.5.5.64,<4.8.0
Pillow>=8.4.0,<10.0.0
PyAutoGUI>=0.9.53,<0.10.0
numpy>=1.19.5,<1.22.0

# System utilities (Python 3.8 compatible)
psutil>=5.8.0,<6.0.0

# Network requests (Python 3.8 compatible)
requests>=2.25.1,<3.0.0

# Alternative keylogger library (easier installation)
keyboard>=1.13.0

# Alternative to some opencv functionality  
mss>=6.1.0  # For screenshots without opencv dependency

# Alternative to PyAutoGUI for some functionality
pygetwindow>=0.0.9  # For window management